requirements.txt
# Core ML libraries
mlflow>=3.0.0
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.2
xgboost==2.0.3

# Data handling
seaborn==0.13.0

# Code formatting
black==23.12.1
isort==5.13.2

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
